@@ -14,8 +14,8 @@
     "prop-types": "^15.6.2",
     "random-int": "^1.0.0",
     "randomcolor": "^0.5.3",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
     "react-motion": "^0.5.2",
     "react-portal": "^4.1.5",
     "react-treeview": "^0.4.7",
