@@ -17,7 +17,9 @@
   "main": "lib/index.js",
   "scripts": {
     "start": "NODE_ENV=development babel src -w -d lib",
-    "build": "NODE_ENV=production babel src -d lib",
+    "build:lib": "NODE_ENV=production babel src -d lib",
+    "build:flow": "NODE_ENV=production flow-copy-source -v src lib",
+    "build": "npm run build:lib && npm run build:flow",
     "lint": "NODE_ENV=development eslint src/**",
     "format": "prettier --trailing-comma es5 --no-semi --single-quote --write \"src/**/*.js\"",
     "test": "cd test && npm install && bash test.sh",
@@ -27,13 +29,11 @@
   },
   "dependencies": {
     "fbjs": "^1.0.0",
-    "lodash.isequal": "^4.5.0",
     "paper": "^0.11.8",
-    "prop-types": "^15.6.2",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-reconciler": "^0.15.0",
-    "scheduler": "^0.9.0"
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
+    "react-reconciler": "^0.16.0",
+    "scheduler": "^0.10.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.1.2",
@@ -43,14 +43,14 @@
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
-    "babel-eslint": "^9.0.0",
-    "babel-preset-env": "^1.7.0",
+    "babel-eslint": "=9.0.0",
     "eslint": "=5.6.0",
     "eslint-plugin-flowtype": "^3.0.0",
     "eslint-plugin-import": "^2.12.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-react": "^7.8.2",
-    "flow-bin": "^0.83.0",
+    "flow-bin": "^0.84.0",
+    "flow-copy-source": "^2.0.2",
     "prettier": "^1.13.3",
     "semantic-release": "^15.10.5",
     "travis-deploy-once": "^5.0.0"
