@@ -3,48 +3,54 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/runtime": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
     "autoprefixer": "7.1.6",
-    "babel-core": "6.26.0",
-    "babel-eslint": "7.2.3",
-    "babel-jest": "20.0.3",
-    "babel-loader": "7.1.2",
-    "babel-preset-react-app": "^3.1.1",
-    "babel-runtime": "6.26.0",
-    "case-sensitive-paths-webpack-plugin": "2.1.1",
-    "chalk": "1.1.3",
-    "css-loader": "0.28.7",
-    "dotenv": "4.0.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "9.0.0",
+    "babel-jest": "^23.4.2",
+    "babel-loader": "^8.0.0",
+    "babel-preset-react-app": "5.0.4",
+    "case-sensitive-paths-webpack-plugin": "2.1.2",
+    "chalk": "2.4.1",
+    "chromedriver": "^2.35.0",
+    "css-loader": "1.0.0",
+    "dotenv": "6.1.0",
     "dotenv-expand": "4.2.0",
-    "eslint": "4.10.0",
-    "eslint-config-react-app": "^2.1.0",
-    "eslint-loader": "1.9.0",
-    "eslint-plugin-flowtype": "2.39.1",
-    "eslint-plugin-import": "2.8.0",
-    "eslint-plugin-jsx-a11y": "5.1.1",
-    "eslint-plugin-react": "7.4.0",
+    "eslint": "5.7.0",
+    "eslint-config-react-app": "3.0.4",
+    "eslint-loader": "2.1.1",
+    "eslint-plugin-flowtype": "2.0.0",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-jsx-a11y": "6.1.2",
+    "eslint-plugin-react": "7.11.1",
     "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "1.1.5",
-    "fs-extra": "3.0.1",
-    "html-webpack-plugin": "2.29.0",
-    "jest": "20.0.4",
+    "file-loader": "2.0.0",
+    "fs-extra": "7.0.0",
+    "html-webpack-plugin": "3.2.0",
+    "jest": "23.6.0",
     "lodash": "^4.17.5",
+    "nightwatch": "^0.9.21",
+    "node-resemble-js": "^0.2.0",
     "object-assign": "4.1.1",
-    "postcss-flexbugs-fixes": "3.2.0",
-    "postcss-loader": "2.0.8",
-    "promise": "8.0.1",
-    "query-string": "^5.1.0",
+    "postcss-flexbugs-fixes": "4.1.0",
+    "postcss-loader": "3.0.0",
+    "promise": "8.0.2",
+    "query-string": "6.2.0",
     "raf": "3.4.0",
     "react": "^16.2.0",
-    "react-dev-utils": "^5.0.0",
+    "react-dev-utils": "6.0.5",
     "react-dom": "^16.2.0",
     "react-router-dom": "^4.2.2",
-    "style-loader": "0.19.0",
-    "sw-precache-webpack-plugin": "0.11.4",
-    "url-loader": "0.6.2",
-    "webpack": "3.8.1",
-    "webpack-dev-server": "2.9.4",
-    "webpack-manifest-plugin": "1.3.2",
-    "whatwg-fetch": "2.0.3"
+    "selenium-server-standalone-jar": "^3.9.1",
+    "style-loader": "0.23.1",
+    "sw-precache-webpack-plugin": "0.11.5",
+    "url-loader": "1.1.2",
+    "webpack": "^4.23.1",
+    "webpack-manifest-plugin": "2.0.4",
+    "webpack-dev-server": "3.1.10",
+    "whatwg-fetch": "3.0.0"
   },
   "scripts": {
     "start": "node scripts/start.js",
@@ -92,11 +98,5 @@
   },
   "eslintConfig": {
     "extends": "react-app"
-  },
-  "devDependencies": {
-    "chromedriver": "^2.35.0",
-    "nightwatch": "^0.9.19",
-    "node-resemble-js": "^0.2.0",
-    "selenium-server-standalone-jar": "^3.9.1"
   }
 }
