@@ -3,73 +3,57 @@
 import React, { Component } from 'react'
 import { PaperScope, Size } from 'paper/dist/paper-core'
 
-import type { Node } from 'react'
+import type { Node, ElementRef } from 'react'
 import type { FiberRoot } from 'react-reconciler'
 
 import PaperRenderer from './PaperRenderer'
 
+type CanvasRef = {
+  current: ElementRef<any> | null,
+}
+
 type Props = {
-  activeLayer?: ?number,
-  activeTool?: ?string,
-  children: Node,
+  children?: Node,
   width: number,
   height: number,
   settings?: Object,
 }
 
 export default class View extends Component<Props> {
-  paper: PaperScope
-  canvas: HTMLElement
+  canvas: CanvasRef
+  scope: PaperScope
   mountNode: FiberRoot
 
-  componentDidMount() {
-    console.log('mount')
-    const {
-      activeLayer,
-      activeTool,
-      children,
-      width,
-      height,
-      settings,
-    } = this.props
+  constructor(props: Props) {
+    super(props)
+    this.canvas = React.createRef()
+  }
 
-    this.paper = new PaperScope()
-    this.paper.setup(this.canvas)
+  componentDidMount() {
+    const { children, width, height, settings } = this.props
+
+    this.scope = new PaperScope()
+    this.scope.setup(this.canvas.current)
 
     if (settings) {
       for (let key of Object.keys(settings)) {
-        this.paper.settings[key] = settings[key]
+        this.scope.settings[key] = settings[key]
       }
     }
 
-    const { project, tools, view } = this.paper
+    this.scope.view.viewSize = new Size(width, height)
 
-    view.viewSize = new Size(width, height)
-
-    this.mountNode = PaperRenderer.createContainer(this.paper)
+    this.mountNode = PaperRenderer.createContainer(this.scope)
 
     PaperRenderer.updateContainer(children, this.mountNode, this)
-
-    // initial active layer
-    if (typeof activeLayer === 'number') {
-      const layer = project.layers.find(l => l.data.id === activeLayer)
-      if (layer) layer.activate()
-    }
-
-    // initial active tool
-    if (typeof activeTool === 'string') {
-      const tool = tools.find(t => t.name === activeTool)
-      if (tool) tool.activate()
-    }
   }
 
   componentDidUpdate(prevProps: Props) {
     const { children, width, height } = this.props
-    const { view } = this.paper
+    const { view } = this.scope
 
     PaperRenderer.updateContainer(children, this.mountNode, this)
 
-    // size has changed, update center
     if (width !== prevProps.width || height !== prevProps.height) {
       const prevCenter = view.center
       view.viewSize = new Size(width, height)
@@ -81,25 +65,15 @@ export default class View extends Component<Props> {
     PaperRenderer.updateContainer(null, this.mountNode, this)
   }
 
-  canvasRef = (ref: ?HTMLElement) => {
-    if (ref) this.canvas = ref
-  }
-
   render() {
-    PaperRenderer.injectIntoDevTools({
-      bundleType: process.env.NODE_ENV === 'production' ? 0 : 1,
-      version: '0.10.0',
-      rendererPackageName: 'react-paper-bindings',
-      findHostInstanceByFiber: PaperRenderer.findHostInstance,
-    })
-    const {
-      activeLayer,
-      activeTool,
-      children,
-      width,
-      height,
-      ...other
-    } = this.props
-    return <canvas {...other} ref={this.canvasRef} />
+    const { children, width, height, ...other } = this.props
+    return <canvas {...other} ref={this.canvas} />
   }
 }
+
+PaperRenderer.injectIntoDevTools({
+  findFiberByHostInstance: () => null,
+  bundleType: process.env.NODE_ENV === 'production' ? 0 : 1,
+  version: '16.5.2',
+  rendererPackageName: 'react-paper-bindings',
+})
